+++ R/transf.R
@@ -16,7 +16,7 @@
 #' @examples
 #'
 #' file <- system.file("extdata", "lanjaron.gpkg", package = "satres")
-#' lanjaron <- sf::st_read(file, layer = "lanjaron_bbox")
+#' lanjaron <- sf::st_read(file, layer = "lanjaron_bbox", quiet = TRUE)
 #'
 #' esa <- system.file("extdata", "esa", package = "satres")
 #' sr <- satres(dir = esa) |>