+++ NEWS.md
@@ -1,4 +1,6 @@
 # satres (development version)
+* Functions to transform an object: `clip_bands()`.
+* Functions to query an object: `get_spectral_resolution()`, `get_all_names()`, `get_band_names()`.
 * Fix typos in the documentation. 
 
 # satres 1.0.0